Transactions

+++ shared/src/commonMain/kotlin/com/victor/restart/core/data/transaction/TransactionRequest.kt
@@ -1,10 +1,11 @@
 package com.victor.restart.core.data.transaction
 
+import com.victor.restart.core.utils.QuotedDoubleSerializer
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class TransactionRequest(
     val categoryId: Long,
-    val amount: Int,
+    val amount: Double,
     val description: String? = null,
 )
